requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
firebase-admin
slowapi
google-cloud-firestore
pytz